Refactored user_messages to be top level
