+++ gitgud/skills/intro/__init__.py
@@ -3,8 +3,8 @@ import subprocess
 from gitgud.skills.util import Skill
 from gitgud.skills.level_builder import BasicLevel
 
-from gitgud.skills.user_messages import simulate_command
-from gitgud.skills.user_messages import default_fail_no_reset
+from gitgud.user_messages import simulate_command
+from gitgud.user_messages import default_fail_no_reset
 
 from gitgud import operations
